#201 changed Test to IT + forced new isolated app to be created and later deleted
