+++ examples/quickstart/pom.xml
@@ -7,7 +7,7 @@
     <parent>
         <groupId>com.stormpath.sdk</groupId>
         <artifactId>stormpath-sdk-root</artifactId>
-        <version>1.0.RC4.3-SNAPSHOT</version>
+        <version>1.0.RC4.6-SNAPSHOT</version>
         <relativePath>../../pom.xml</relativePath>
     </parent>
